SpriteChanges, enemy changes
I added in the fire ball sprite images to the character, added the big character sprites to the game by changing the animation when you get the growth potion. I also changed the colliders of the enemy and made it so i can destroy them by stomping on them like in mario.